1.0 - Rich Clune Update
It works hard, and it works.

Had to delete and reupload entire repository due to keys being accidentily committed. Anyways, onto the main stuff.

get_Schedule
- Has to be updated yearly with the new BASE_SEASON_ID which goes up by 1-5 depending on whatever the AHL is doing and their internal logic that year.
- Will build the URL for the ICS file that you can download of the entire season.
- Downloads the ICS file
- Parses and pushes game_data to SQLite database.

main.py
- uses User classes that I like to use in all my projects. Twitter_user was "udpated" to reflect that but you can't see that here.

- The primary thing that the program does is check if it is a gameday, and if it is, it search for a lineup image from @TorontoMarlies and parse that lineup to a reddit post with other informaiton such as streaming info.

This program is a bit of a mess right now, but trust me, it was messier. I'll be working on commenting everything and refining & distilling everything as the 2019 season draws closer.